+++ binstar_client/commands/upload.py
@@ -7,16 +7,14 @@
 ##### See Also
 
   * [Uploading a Conda Package](https://docs.anaconda.com/anaconda-repository/user-guide/tasks/pkgs/use-pkg-managers/#uploading-a-conda-package)
-  * [Uploading a PyPI Package](https://docs.anaconda.com/anaconda-repository/user-guide/tasks/pkgs/use-pkg-managers/#uploading-pypi-packages)
+  * [Uploading a Standard Python Package](https://docs.anaconda.com/anaconda-repository/user-guide/tasks/pkgs/use-pkg-managers/#uploading-pypi-packages)
 
 """
 from __future__ import unicode_literals
 
 import argparse
-import tempfile
 import logging
 import os
-import subprocess
 
 from glob import glob
 from os.path import exists
@@ -27,7 +25,7 @@ from six.moves import input
 
 from binstar_client import errors
 from binstar_client.utils import bool_input, DEFAULT_CONFIG, get_config, get_server_api, upload_print_callback
-from binstar_client.utils.config import PACKAGE_TYPES
+from binstar_client.utils.config import PackageType
 from binstar_client.utils.projects import upload_project
 from binstar_client.utils.detect import detect_package_type, get_attrs
 
@@ -36,7 +34,7 @@ logger = logging.getLogger('binstar.upload')
 
 
 def verbose_package_type(pkg_type, lowercase=True):
-    verbose_type = PACKAGE_TYPES.get(pkg_type, 'unknown')
+    verbose_type = pkg_type.label()
     if lowercase:
         verbose_type = verbose_type.lower()
     return verbose_type
@@ -71,23 +69,23 @@ def determine_package_type(filename, args):
     -t/--package-type argument
     """
     if args.package_type:
-        package_type = args.package_type
-    else:
-        logger.info('Detecting file type...')
+        return PackageType(args.package_type)
 
-        package_type = detect_package_type(filename)
+    logger.info('Detecting file type...')
+    package_type = detect_package_type(filename)
 
-        if package_type is None:
-            message = 'Could not detect package type of file %r please specify package type with option --package-type' % filename
-            logger.error(message)
-            raise errors.BinstarError(message)
+    if package_type is None:
+        message = 'Could not detect package type of file %r please ' \
+                  'specify package type with option --package-type' % filename
+        logger.error(message)
+        raise errors.BinstarError(message)
 
-        logger.info('File type is "%s"', package_type)
+    logger.info('File type is "%s"', package_type.label())
 
     return package_type
 
 
-def get_package_name(args, package_attrs, filename, package_type):
+def get_package_name(args, package_attrs, package_type):
     if args.package:
         if 'name' in package_attrs and package_attrs['name'].lower() != args.package.lower():
             msg = 'Package name on the command line " {}" does not match the package name in the file "{}"'.format(
@@ -98,7 +96,8 @@ def get_package_name(args, package_attrs, filename, package_type):
         package_name = args.package
     else:
         if 'name' not in package_attrs:
-            message = "Could not detect package name for package type %s, please use the --package option" % (package_type,)
+            message = "Could not detect package name for package type {}, please use the --package option".format(
+                package_type.label())
             logger.error(message)
             raise errors.BinstarError(message)
         package_name = package_attrs['name']
@@ -111,7 +110,8 @@ def get_version(args, release_attrs, package_type):
         version = args.version
     else:
         if 'version' not in release_attrs:
-            message = "Could not detect package version for package type %s, please use the --version option" % (package_type,)
+            message = "Could not detect package version for package type \"{}\", " \
+                      "please use the --version option".format(package_type.label())
             logger.error(message)
             raise errors.BinstarError(message)
         version = release_attrs['version']
@@ -209,16 +209,16 @@ def upload_package(filename, package_type, aserver_api, username, args):
     if args.description:
         release_attrs['description'] = args.description
 
-    package_name = get_package_name(args, package_attrs, filename, package_type)
+    package_name = get_package_name(args, package_attrs, package_type)
     version = get_version(args, release_attrs, package_type)
 
     logger.info('Creating package "%s"', package_name)
 
     package = add_package(aserver_api, args, username, package_name, package_attrs, package_type)
-    package_types = package.get('package_types', [])
+    package_types = [PackageType(pkg_type) for pkg_type in package.get('package_types', [])]
 
     allowed_package_types = set(package_types)
-    for group in [{'conda', 'pypi'}]:
+    for group in [{PackageType.CONDA, PackageType.STANDARD_PYTHON}]:
         if allowed_package_types & group:
             allowed_package_types.update(group)
 
@@ -260,28 +260,6 @@ def upload_package(filename, package_type, aserver_api, username, args):
     return [package_name, upload_info]
 
 
-def get_convert_files(files):
-    tmpdir = tempfile.mkdtemp()
-
-    for filepath in files:
-        logger.info('Running conda convert on "%s"', filepath)
-        process = subprocess.Popen(
-            ['conda-convert', '-p', 'all', filepath, '-o', tmpdir],
-            stdout=subprocess.PIPE, stderr=subprocess.PIPE
-        )
-        stdout, stderr = process.communicate()
-
-        if stderr:
-            logger.warning('Couldn\'t generate platform packages for %s: %s', filepath, stderr)
-
-    result = []
-    for path, dirs, files in os.walk(tmpdir):
-        for filename in files:
-            result.append(os.path.join(path, filename))
-
-    return result
-
-
 def main(args):
     config = get_config(site=args.site)
 
@@ -313,10 +291,7 @@ def main(args):
     uploaded_projects = []
 
     # Flatten file list because of 'windows_glob' function
-    files = [f for fglob in args.files for f in fglob]
-
-    if args.all:
-        files += get_convert_files(files)
+    files = list(set(f for fglob in args.files for f in fglob))
 
     for filename in files:
         if not exists(filename):
@@ -328,10 +303,10 @@ def main(args):
 
         package_type = determine_package_type(filename, args)
 
-        if package_type == 'project':
+        if package_type is PackageType.PROJECT:
             uploaded_projects.append(upload_project(filename, args, username))
         else:
-            if package_type == 'ipynb' and not args.mode == 'force':
+            if package_type is PackageType.NOTEBOOK and not args.mode == 'force':
                 try:
                     nbformat.read(open(filename), nbformat.NO_CONVERT)
                 except Exception as error:
@@ -349,9 +324,9 @@ def main(args):
             if package_info is not None and len(package_info) == 2:
                 _package, _upload_info = package_info
                 if _upload_info:
-                    uploaded_packages.append(package_info)
+                    uploaded_packages.append((_package, _upload_info, package_type))
 
-    for package, upload_info in uploaded_packages:
+    for package, upload_info, package_type in uploaded_packages:
         package_url = upload_info.get('url', 'https://anaconda.org/%s/%s' % (username, package))
         logger.info("{} located at:\n{}\n".format(verbose_package_type(package_type), package_url))
 
@@ -359,67 +334,131 @@ def main(args):
         logger.info("Project {} uploaded to {}.\n".format(project_name, url))
 
 
-def windows_glob(item):
-    if os.name == 'nt' and '*' in item:
+def pathname_list(item):
+    if (os.name == 'nt') and any(character in {'*', '?'} for character in item):
         return glob(item)
-    else:
-        return [item]
+    return [item]
 
 
 def add_parser(subparsers):
     description = 'Upload packages to your Anaconda repository'
-    parser = subparsers.add_parser('upload',
-                                   formatter_class=argparse.RawDescriptionHelpFormatter,
-                                   help=description, description=description,
-                                   epilog=__doc__)
+    parser = subparsers.add_parser(
+        'upload',
+        formatter_class=argparse.RawDescriptionHelpFormatter,
+        help=description, description=description,
+        epilog=__doc__,
+    )
 
-    parser.add_argument('files', nargs='+', help='Distributions to upload', default=[], type=windows_glob)
+    parser.add_argument('files', nargs='+', help='Distributions to upload', default=[], type=pathname_list)
 
     label_help = (
         '{deprecation}Add this file to a specific {label}. '
         'Warning: if the file {label}s do not include "main", '
         'the file will not show up in your user {label}')
 
-    parser.add_argument('-c', '--channel', action='append', default=[], dest='labels',
-                        help=label_help.format(deprecation='[DEPRECATED]\n', label='channel'),
-                        metavar='CHANNELS')
-    parser.add_argument('-l', '--label', action='append', dest='labels',
-                        help=label_help.format(deprecation='', label='label'))
-    parser.add_argument('--no-progress', help="Don't show upload progress", action='store_true')
-    parser.add_argument('-u', '--user', help='User account or Organization, defaults to the current user')
-    parser.add_argument('--all', help='Use conda convert to generate packages for all platforms and upload them',
-                        action='store_true')
+    parser.add_argument(
+        '-c', '--channel',
+        action='append',
+        default=[],
+        dest='labels',
+        help=label_help.format(deprecation='[DEPRECATED]\n', label='channel'),
+        metavar='CHANNELS',
+    )
+    parser.add_argument(
+        '-l', '--label',
+        action='append',
+        dest='labels',
+        help=label_help.format(deprecation='', label='label'),
+    )
+    parser.add_argument(
+        '--no-progress',
+        help="Don't show upload progress",
+        action='store_true',
+    )
+    parser.add_argument(
+        '-u', '--user',
+        help='User account or Organization, defaults to the current user',
+    )
 
     mgroup = parser.add_argument_group('metadata options')
-    mgroup.add_argument('-p', '--package', help='Defaults to the package name in the uploaded file')
-    mgroup.add_argument('-v', '--version', help='Defaults to the package version in the uploaded file')
-    mgroup.add_argument('-s', '--summary', help='Set the summary of the package')
-    # To preserve current behavior
-    pkgs = PACKAGE_TYPES.copy()
-    pkgs.pop('conda')
-    pkgs.pop('pypi')
-    pkg_types = ', '.join(list(pkgs.keys()))
-    mgroup.add_argument('-t', '--package-type', help='Set the package type [{0}]. Defaults to autodetect'.format(pkg_types))
-    mgroup.add_argument('-d', '--description', help='description of the file(s)')
-    mgroup.add_argument('--thumbnail', help='Notebook\'s thumbnail image')
-    mgroup.add_argument('--private', help="Create the package with private access", action='store_true')
+    mgroup.add_argument(
+        '-p', '--package',
+        help='Defaults to the package name in the uploaded file',
+    )
+    mgroup.add_argument(
+        '-v', '--version',
+        help='Defaults to the package version in the uploaded file',
+    )
+    mgroup.add_argument(
+        '-s', '--summary',
+        help='Set the summary of the package',
+    )
+    mgroup.add_argument(
+        '-t', '--package-type',
+        help='Set the package type. Defaults to autodetect',
+    )
+    mgroup.add_argument(
+        '-d', '--description',
+        help='description of the file(s)',
+    )
+    mgroup.add_argument(
+        '--thumbnail',
+        help="Notebook's thumbnail image",
+    )
+    mgroup.add_argument(
+        '--private',
+        help='Create the package with private access',
+        action='store_true',
+    )
 
     register_group = parser.add_mutually_exclusive_group()
-    register_group.add_argument("--no-register", dest="auto_register", action="store_false",
-                        help='Don\'t create a new package namespace if it does not exist')
-    register_group.add_argument("--register", dest="auto_register", action="store_true",
-                        help='Create a new package namespace if it does not exist')
+    register_group.add_argument(
+        '--no-register',
+        dest='auto_register',
+        action='store_false',
+        help="Don't create a new package namespace if it does not exist",
+    )
+    register_group.add_argument(
+        '--register',
+        dest='auto_register',
+        action='store_true',
+        help='Create a new package namespace if it does not exist',
+    )
+
     parser.set_defaults(auto_register=DEFAULT_CONFIG.get('auto_register', True))
-    parser.add_argument('--build-id', help='Anaconda repository Build ID (internal only)')
+    parser.add_argument(
+        '--build-id',
+        help='Anaconda repository Build ID (internal only)',
+    )
 
     group = parser.add_mutually_exclusive_group()
-    group.add_argument('-i', '--interactive', action='store_const', help='Run an interactive prompt if any packages are missing',
-                        dest='mode', const='interactive')
-    group.add_argument('-f', '--fail', help='Fail if a package or release does not exist (default)',
-                                        action='store_const', dest='mode', const='fail')
-    group.add_argument('--force', help='Force a package upload regardless of errors',
-                                        action='store_const', dest='mode', const='force')
-    group.add_argument('--skip-existing', help='Skip errors on package batch upload if it already exists',
-                                        action='store_const', dest='mode', const='skip')
+    group.add_argument(
+        '-i', '--interactive',
+        action='store_const',
+        help='Run an interactive prompt if any packages are missing',
+        dest='mode',
+        const='interactive',
+    )
+    group.add_argument(
+        '-f', '--fail',
+        help='Fail if a package or release does not exist (default)',
+        action='store_const',
+        dest='mode',
+        const='fail',
+    )
+    group.add_argument(
+        '--force',
+        help='Force a package upload regardless of errors',
+        action='store_const',
+        dest='mode',
+        const='force',
+    )
+    group.add_argument(
+        '--skip-existing',
+        help='Skip errors on package batch upload if it already exists',
+        action='store_const',
+        dest='mode',
+        const='skip',
+    )
 
     parser.set_defaults(main=main)