+++ binstar_client/tests/test_config.py
@@ -1,11 +1,8 @@
 from __future__ import unicode_literals
 
-import os
 import shutil
 import tempfile
-
 from os.path import join, exists
-from operator import delitem
 
 import mock