+++ binstar_client/utils/appdirs.py
@@ -1,4 +1,5 @@
-#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
 # Copyright (c) 2005-2010 ActiveState Software Inc.
 
 """Utilities for determining application-specific dirs.
@@ -23,9 +24,9 @@ PY3 = sys.version_info[0] == 3
 if PY3:
     unicode = str
 
+
 class AppDirsError(Exception):
     pass
-
 
 
 def user_data_dir(appname, appauthor=None, version=None, roaming=False):
@@ -66,13 +67,9 @@ def user_data_dir(appname, appauthor=None, version=None, roaming=False):
         const = roaming and "CSIDL_APPDATA" or "CSIDL_LOCAL_APPDATA"
         path = os.path.join(_get_win_folder(const), appauthor, appname)
     elif sys.platform == 'darwin':
-        path = os.path.join(
-            os.path.expanduser('~/Library/Application Support/'),
-            appname)
+        path = os.path.join(os.path.expanduser('~/Library/Application Support/'), appname)
     else:
-        path = os.path.join(
-            os.getenv('XDG_CONFIG_HOME', os.path.expanduser("~/.config")),
-            appname.lower())
+        path = os.path.join(os.getenv('XDG_CONFIG_HOME', os.path.expanduser("~/.config")), appname.lower())
     if version:
         path = os.path.join(path, version)
     return path
@@ -104,15 +101,11 @@ def site_data_dir(appname, appauthor=None, version=None):
     if sys.platform.startswith("win"):
         if appauthor is None:
             raise AppDirsError("must specify 'appauthor' on Windows")
-        path = os.path.join(_get_win_folder("CSIDL_COMMON_APPDATA"),
-                            appauthor, appname)
+        path = os.path.join(_get_win_folder("CSIDL_COMMON_APPDATA"), appauthor, appname)
     elif sys.platform == 'darwin':
-        path = os.path.join(
-            os.path.expanduser('/Library/Application Support'),
-            appname)
+        path = os.path.join(os.path.expanduser('/Library/Application Support'), appname)
     else:
-        # XDG default for $XDG_CONFIG_DIRS[0]. Perhaps should actually
-        # *use* that envvar, if defined.
+        # XDG default for $XDG_CONFIG_DIRS[0]. Perhaps should actually *use* that envvar, if defined.
         path = "/etc/xdg/" + appname.lower()
     if version:
         path = os.path.join(path, version)
@@ -152,21 +145,17 @@ def user_cache_dir(appname, appauthor=None, version=None, opinion=True):
     if sys.platform.startswith("win"):
         if appauthor is None:
             raise AppDirsError("must specify 'appauthor' on Windows")
-        path = os.path.join(_get_win_folder("CSIDL_LOCAL_APPDATA"),
-                            appauthor, appname)
+        path = os.path.join(_get_win_folder("CSIDL_LOCAL_APPDATA"), appauthor, appname)
         if opinion:
             path = os.path.join(path, "Cache")
     elif sys.platform == 'darwin':
-        path = os.path.join(
-            os.path.expanduser('~/Library/Caches'),
-            appname)
+        path = os.path.join(os.path.expanduser('~/Library/Caches'), appname)
     else:
-        path = os.path.join(
-            os.getenv('XDG_CACHE_HOME', os.path.expanduser('~/.cache')),
-            appname.lower())
+        path = os.path.join(os.getenv('XDG_CACHE_HOME', os.path.expanduser('~/.cache')), appname.lower())
     if version:
         path = os.path.join(path, version)
     return path
+
 
 def user_log_dir(appname, appauthor=None, version=None, opinion=True):
     r"""Return full path to the user-specific log dir for this application.
@@ -198,32 +187,37 @@ def user_log_dir(appname, appauthor=None, version=None, opinion=True):
     This can be disabled with the `opinion=False` option.
     """
     if sys.platform == "darwin":
-        path = os.path.join(
-            os.path.expanduser('~/Library/Logs'),
-            appname)
+        path = os.path.join(os.path.expanduser('~/Library/Logs'), appname)
     elif sys.platform == "win32":
-        path = user_data_dir(appname, appauthor, version); version = False
+        path = user_data_dir(appname, appauthor, version)
+        version = False
         if opinion:
             path = os.path.join(path, "Logs")
     else:
-        path = user_cache_dir(appname, appauthor, version); version = False
+        path = user_cache_dir(appname, appauthor, version)
+        version = False
         if opinion:
             path = os.path.join(path, "log")
     if version:
         path = os.path.join(path, version)
     return path
 
+
 class EnvAppDirs(object):
+
     def __init__(self, appname, appauthor, root_path):
         self.appname = appname
         self.appauthor = appauthor
         self.root_path = root_path
+
     @property
     def user_data_dir(self):
         return os.path.join(self.root_path, 'data')
+
     @property
     def site_data_dir(self):
         return os.path.join(self.root_path, 'data')
+
     @property
     def user_cache_dir(self):
         return os.path.join(self.root_path, 'cache')
@@ -235,37 +229,36 @@ class EnvAppDirs(object):
 
 class AppDirs(object):
     """Convenience wrapper for getting application dirs."""
+
     def __init__(self, appname, appauthor, version=None, roaming=False):
         self.appname = appname
         self.appauthor = appauthor
         self.version = version
         self.roaming = roaming
+
     @property
     def user_data_dir(self):
-        return user_data_dir(self.appname, self.appauthor,
-            version=self.version, roaming=self.roaming)
+        return user_data_dir(self.appname, self.appauthor, version=self.version, roaming=self.roaming)
+
     @property
     def site_data_dir(self):
-        return site_data_dir(self.appname, self.appauthor,
-            version=self.version)
+        return site_data_dir(self.appname, self.appauthor, version=self.version)
+
     @property
     def user_cache_dir(self):
-        return user_cache_dir(self.appname, self.appauthor,
-            version=self.version)
+        return user_cache_dir(self.appname, self.appauthor, version=self.version)
+
     @property
     def user_log_dir(self):
-        return user_log_dir(self.appname, self.appauthor,
-            version=self.version)
+        return user_log_dir(self.appname, self.appauthor, version=self.version)
 
 
-
-
-#---- internal support stuff
+# ---- internal support stuff
 
 def _get_win_folder_from_registry(csidl_name):
-    """This is a fallback technique at best. I'm not sure if using the
-    registry for this guarantees us the correct answer for all CSIDL_*
-    names.
+    """
+    This is a fallback technique at best. I'm not sure if using the registry for this guarantees us the correct answer
+    for all CSIDL_* names.
     """
     import _winreg
 
@@ -275,17 +268,20 @@ def _get_win_folder_from_registry(csidl_name):
         "CSIDL_LOCAL_APPDATA": "Local AppData",
     }[csidl_name]
 
-    key = _winreg.OpenKey(_winreg.HKEY_CURRENT_USER,
-        r"Software\Microsoft\Windows\CurrentVersion\Explorer\Shell Folders")
+    key = _winreg.OpenKey(
+        _winreg.HKEY_CURRENT_USER,
+        r'Software\Microsoft\Windows\CurrentVersion\Explorer\Shell Folders'
+    )
     dir, type = _winreg.QueryValueEx(key, shell_folder_name)
     return dir
+
 
 def _get_win_folder_with_pywin32(csidl_name):
     from win32com.shell import shellcon, shell
     dir = shell.SHGetFolderPath(0, getattr(shellcon, csidl_name), 0, 0)
-    # Try to make this a unicode path because SHGetFolderPath does
-    # not return unicode strings when there is unicode data in the
-    # path.
+
+    # Try to make this a unicode path because SHGetFolderPath does not return unicode strings when there is unicode data
+    # in the path.
     try:
         dir = unicode(dir)
 
@@ -305,6 +301,7 @@ def _get_win_folder_with_pywin32(csidl_name):
     except UnicodeError:
         pass
     return dir
+
 
 def _get_win_folder_with_ctypes(csidl_name):
     import ctypes
@@ -332,6 +329,7 @@ def _get_win_folder_with_ctypes(csidl_name):
 
     return buf.value
 
+
 if sys.platform == "win32":
     try:
         import win32com.shell
@@ -344,8 +342,7 @@ if sys.platform == "win32":
             _get_win_folder = _get_win_folder_from_registry
 
 
-
-#---- self test code
+# ---- self test code
 
 if __name__ == "__main__":
     appname = "MyApp"