+++ binstar_client/utils/pprint.py
@@ -3,14 +3,21 @@ Created on Aug 8, 2013
 
 @author: sean
 '''
+
 from __future__ import unicode_literals
+from binstar_client.utils import config
 from dateutil.parser import parse as parse_date
 import logging
 
 logger = logging.getLogger('binstar.pprint')
 
-fmt_access = '     %(full_name)-25s | %(latest_version)8s | %(access)-12s | %(package_types)-15s | %(conda_platforms)-15s | %(builds)-10s'
-fmt_no_access = '     %(full_name)-25s | %(latest_version)8s | %(package_types)-15s | %(conda_platforms)-15s | %(builds)-10s'
+fmt_access = (
+    '     %(full_name)-32s | %(latest_version)8s | %(access)-12s | %(package_types)-17s | %(conda_platforms)-15s | '
+    '%(builds)-10s'
+)
+fmt_no_access = (
+    '     %(full_name)-32s | %(latest_version)8s | %(package_types)-17s | %(conda_platforms)-15s | %(builds)-10s'
+)
 
 
 def pprint_orgs(orgs):
@@ -37,22 +44,47 @@ def pprint_package_header(access=True, revisions=False):
     logger.info(fmt % package_header)
 
 
+def format_package_type(value):
+    value = str(value)
+    try:
+        return config.PackageType(value).label()
+    except ValueError:
+        return value
+
+
 def pprint_package(package, access=True, full_name=True, revision=False):
     package = package.copy()
 
-    package['access'] = 'published' if package.get('published') else 'public' if package['public'] else 'private'
+    if package.get('published'):
+        package['access'] = 'published'
+    elif package['public']:
+        package['access'] = 'public'
+    else:
+        package['access'] = 'private'
 
     if package.get('conda_platforms'):
-        package['conda_platforms'] = ', '.join(str(x) for x in package['conda_platforms'] if x is not None)
+        package['conda_platforms'] = ', '.join(
+            str(x)
+            for x in package['conda_platforms']
+            if x is not None
+        )
 
     if not full_name:
         package['full_name'] = package['name']
 
     if package.get('package_types'):
-        package['package_types'] = ', '.join(str(x) for x in package['package_types'] if x is not None)
+        package['package_types'] = ', '.join(
+            format_package_type(x)
+            for x in package['package_types']
+            if x is not None
+        )
 
     if package.get('builds'):
-        package['builds'] = ', '.join(str(x) for x in package['builds'] if x is not None)
+        package['builds'] = ', '.join(
+            str(x)
+            for x in package['builds']
+            if x is not None
+        )
     else:
         package['builds'] = ''
 
@@ -77,11 +109,11 @@ def pprint_packages(packages, access=True, full_name=True, revisions=False):
     pprint_package_header(access, revisions=revisions)
 
     package_header = {
-        'full_name': '-' * 25,
+        'full_name': '-' * 32,
         'access': '-' * 12,
         'latest_version': '-' * 6,
         'conda_platforms': '-' * 15,
-        'package_types': '-' * 15,
+        'package_types': '-' * 17,
         'revision': '-' * 6,
         'builds': '-' * 10
     }